requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Mail==0.10.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.2.1
waitress==3.0.1
